dragondiceCTk/pyproject.toml
[project]
name = "dragondicectk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "customtkinter>=5.2.2",
    "pillow>=11.3.0",
]
